ai-image-detector/requirements.txt
# AI Image Detector - Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core Deep Learning ──
torch>=2.0.0
torchvision>=0.15.0

# ── Web Server ──
flask>=3.0.0
Pillow>=10.0.0

# ── Data Science ──
numpy>=1.24.0
scikit-learn>=1.3.0

# ── Visualization (optional, for evaluation plots) ──
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Dataset Download (optional, for CIFAKE) ──
kaggle>=1.6.0

# ── Utilities ──
tqdm>=4.65.0
